requirements.txt
gym==0.14.0
matplotlib==3.1.1
networkx==2.6.3
neat-python==0.92
numpy==1.21.2
openpyxl==3.0.7
scikit-learn==1.0.2
scipy==1.4.1
torch==1.11.0